requirements.txt
openai>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
lxml>=4.9.3
tenacity>=8.2.3
groq>=0.5.0
orjson>=3.9.5
schedule>=1.2.0
aiohttp>=3.9.0

# NEW additions for Groq + LangChain
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0